+++ src/types/Semigroup.js
@@ -1,4 +1,8 @@
 // @flow
 export type Semigroup<Static, In, OutA, OutB> = {
-  concat: ((In, Static) => OutA, (OutA, Static) => OutB) => (In, Static) => OutB
+  concat: (
+    (In, Static) => OutA,
+    (OutA, Static) => OutB
+  ) => (In, Static) => OutB,
+  concatAll: (...(In, Static) => In) => (In, Static) => In
 };